Remove Timelines contracts.
