+++ mastodon4j-rx/src/main/java/com/sys1yagi/mastodon4j/rx/RxMutes.kt
@@ -1,6 +1,7 @@
 package com.sys1yagi.mastodon4j.rx
 
 import com.sys1yagi.mastodon4j.MastodonClient
+import com.sys1yagi.mastodon4j.api.Pageable
 import com.sys1yagi.mastodon4j.api.Range
 import com.sys1yagi.mastodon4j.api.entity.Account
 import com.sys1yagi.mastodon4j.api.method.Mutes
@@ -9,7 +10,7 @@ import io.reactivex.Single
 class RxMutes(client: MastodonClient) {
     val mutes = Mutes(client)
 
-    fun getMutes(range: Range = Range()): Single<List<Account>> {
+    fun getMutes(range: Range = Range()): Single<Pageable<Account>> {
         return Single.create {
             try {
                 val accounts = mutes.getMutes(range)
