+++ mastodon4j-rx/src/main/java/com/sys1yagi/mastodon4j/rx/RxPublic.kt
@@ -1,10 +1,14 @@
 package com.sys1yagi.mastodon4j.rx
 
 import com.sys1yagi.mastodon4j.MastodonClient
+import com.sys1yagi.mastodon4j.api.Pageable
+import com.sys1yagi.mastodon4j.api.Range
 import com.sys1yagi.mastodon4j.api.entity.Instance
 import com.sys1yagi.mastodon4j.api.entity.Results
+import com.sys1yagi.mastodon4j.api.entity.Status
 import com.sys1yagi.mastodon4j.api.method.Public
 import com.sys1yagi.mastodon4j.rx.extensions.onErrorIfNotDisposed
+import com.sys1yagi.mastodon4j.rx.extensions.single
 import io.reactivex.Single
 
 class RxPublic(client: MastodonClient) {
@@ -29,6 +33,30 @@ class RxPublic(client: MastodonClient) {
             } catch (e: Throwable) {
                 it.onError(e)
             }
+        }
+    }
+
+    fun getLocalPublic(range: Range = Range()): Single<Pageable<Status>> {
+        return single {
+            publicMethod.getLocalPublic(range)
+        }
+    }
+
+    fun getFederatedPublic(range: Range = Range()): Single<Pageable<Status>> {
+        return single {
+            publicMethod.getFederatedPublic(range)
+        }
+    }
+
+    fun getLocalTag(tag: String, range: Range = Range()): Single<Pageable<Status>> {
+        return single {
+            publicMethod.getLocalTag(tag, range)
+        }
+    }
+
+    fun getFederatedTag(tag: String, range: Range = Range()): Single<Pageable<Status>> {
+        return single {
+            publicMethod.getFederatedTag(tag, range)
         }
     }
 }
