+++ mastodon4j-rx/src/test/java/com/sys1yagi/mastodon4j/rx/RxTimelinesTest.kt
@@ -1,7 +1,6 @@
 package com.sys1yagi.mastodon4j.rx
 
 import com.sys1yagi.mastodon4j.rx.testtool.MockClient
-import org.junit.Assert
 import org.junit.Test
 
 class RxTimelinesTest {
@@ -9,8 +8,8 @@ class RxTimelinesTest {
     @Test
     fun getPublic() {
         val client = MockClient.mock("public_timeline.json", 5L, 40L)
-        val timelines = RxTimelines(client)
-        val subscriber = timelines.getLocalPublic().test()
+        val publicMethod = RxPublic(client)
+        val subscriber = publicMethod.getLocalPublic().test()
         subscriber.assertNoErrors()
     }
 }
