+++ mastodon4j/src/main/java/com/sys1yagi/mastodon4j/api/method/Public.kt
@@ -2,9 +2,14 @@ package com.sys1yagi.mastodon4j.api.method
 
 import com.sys1yagi.mastodon4j.MastodonClient
 import com.sys1yagi.mastodon4j.Parameter
+import com.sys1yagi.mastodon4j.api.Pageable
+import com.sys1yagi.mastodon4j.api.Range
 import com.sys1yagi.mastodon4j.api.entity.Instance
 import com.sys1yagi.mastodon4j.api.entity.Results
+import com.sys1yagi.mastodon4j.api.entity.Status
 import com.sys1yagi.mastodon4j.api.exception.Mastodon4jRequestException
+import com.sys1yagi.mastodon4j.extension.genericType
+import com.sys1yagi.mastodon4j.extension.toPageable
 
 class Public(private val client: MastodonClient) {
     /**
@@ -54,4 +59,68 @@ class Public(private val client: MastodonClient) {
             throw Mastodon4jRequestException(response)
         }
     }
+
+    /**
+     *  GET /api/v1/timelines/public
+     * @see https://github.com/tootsuite/documentation/blob/master/Using-the-API/API.md#timelines
+     */
+    private fun getPublic(local: Boolean, range: Range): Pageable<Status> {
+        val parameter = range.toParameter()
+        if (local) {
+            parameter.append("local", local)
+        }
+        val response = client.get(
+                "timelines/public",
+                parameter
+        )
+        if (response.isSuccessful) {
+            val body = response.body().string()
+            return client.getSerializer().fromJson<List<Status>>(
+                    body,
+                    genericType<List<Status>>()
+            ).toPageable(response)
+        } else {
+            throw Mastodon4jRequestException(response)
+        }
+    }
+
+    @JvmOverloads
+    @Throws(Mastodon4jRequestException::class)
+    fun getLocalPublic(range: Range = Range()) = getPublic(true, range)
+
+    @JvmOverloads
+    @Throws(Mastodon4jRequestException::class)
+    fun getFederatedPublic(range: Range = Range()) = getPublic(false, range)
+
+    /**
+     * GET /api/v1/timelines/tag/:tag
+     * @see https://github.com/tootsuite/documentation/blob/master/Using-the-API/API.md#timelines
+     */
+    private fun getTag(tag: String, local: Boolean, range: Range): Pageable<Status> {
+        val parameter = range.toParameter()
+        if (local) {
+            parameter.append("local", local)
+        }
+        val response = client.get(
+                "timelines/tag/$tag",
+                parameter
+        )
+        if (response.isSuccessful) {
+            val body = response.body().string()
+            return client.getSerializer().fromJson<List<Status>>(
+                    body,
+                    genericType<List<Status>>()
+            ).toPageable(response)
+        } else {
+            throw Mastodon4jRequestException(response)
+        }
+    }
+
+    @JvmOverloads
+    @Throws(Mastodon4jRequestException::class)
+    fun getLocalTag(tag: String, range: Range = Range()) = getTag(tag, true, range)
+
+    @JvmOverloads
+    @Throws(Mastodon4jRequestException::class)
+    fun getFederatedTag(tag: String, range: Range = Range()) = getTag(tag, false, range)
 }
