+++ mastodon4j/src/test/java/com/sys1yagi/mastodon4j/api/method/PublicTest.kt
@@ -21,4 +21,27 @@ class PublicTest {
         result.hashtags.size shouldEqualTo 5
         result.hashtags.size shouldEqualTo 5
     }
+
+    @Test
+    fun public() {
+        val client = MockClient.mock("public_timeline.json", maxId = 3L, sinceId = 1L)
+        val publicMethod = Public(client)
+        val statuses = publicMethod.getLocalPublic()
+        statuses.part.size shouldEqualTo 20
+        statuses.link?.let {
+            it.nextPath shouldEqualTo "<https://mstdn.jp/api/v1/timelines/public?limit=20&local=true&max_id=3>"
+            it.maxId shouldEqualTo 3L
+            it.prevPath shouldEqualTo "<https://mstdn.jp/api/v1/timelines/public?limit=20&local=true&since_id=1>"
+            it.sinceId shouldEqualTo 1L
+        }
+    }
+
+    @Test
+    fun tag() {
+        val client = MockClient.mock("tag.json", maxId = 3L, sinceId = 1L)
+        val publicMethod = Public(client)
+        val statuses = publicMethod.getLocalTag("mastodon")
+        statuses.part.size shouldEqualTo 20
+    }
+
 }
