--- sample-kotlin/src/main/java/com/sys1yagi/mastodon4j/sample/AppRegistrationStore.kt
@@ -1,4 +1,0 @@
-package com.sys1yagi.mastodon4j.sample
-
-class AppRegistrationStore(val filePath: String) {
-}
