+++ sample-kotlin/src/main/java/com/sys1yagi/mastodon4j/sample/GetPublicTimelines.kt
@@ -4,16 +4,17 @@ import com.google.gson.Gson
 import com.sys1yagi.mastodon4j.MastodonClient
 import com.sys1yagi.mastodon4j.api.Range
 import com.sys1yagi.mastodon4j.api.exception.Mastodon4jRequestException
+import com.sys1yagi.mastodon4j.api.method.Public
 import com.sys1yagi.mastodon4j.api.method.Timelines
 import okhttp3.OkHttpClient
 
 object GetPublicTimelines {
     @JvmStatic fun main(args: Array<String>) {
         val client = MastodonClient("mstdn.jp", OkHttpClient(), Gson())
-        val timelines = Timelines(client)
+        val publicMethods = Public(client)
 
         try {
-            val statuses = timelines.getLocalPublic(Range())
+            val statuses = publicMethods.getLocalPublic(Range())
             statuses.part.forEach { status ->
                 println("=============")
                 System.out.println(status.account?.displayName)
