+++ sample/src/main/java/com/sys1yagi/mastodon4j/sample/GetTagTimelines.java
@@ -6,6 +6,7 @@ import com.sys1yagi.mastodon4j.api.Pageable;
 import com.sys1yagi.mastodon4j.api.Range;
 import com.sys1yagi.mastodon4j.api.entity.Status;
 import com.sys1yagi.mastodon4j.api.exception.Mastodon4jRequestException;
+import com.sys1yagi.mastodon4j.api.method.Public;
 import com.sys1yagi.mastodon4j.api.method.Timelines;
 import okhttp3.OkHttpClient;
 
@@ -14,10 +15,10 @@ import java.util.List;
 public class GetTagTimelines {
     public static void main(String[] args) {
         MastodonClient client = new MastodonClient("mstdn.jp", new OkHttpClient(), new Gson());
-        Timelines timelines = new Timelines(client);
+        Public publicMethod = new Public(client);
 
         try {
-            Pageable<Status> statuses = timelines.getFederatedTag("mastodon", new Range());
+            Pageable<Status> statuses = publicMethod.getFederatedTag("mastodon", new Range());
             statuses.getPart().forEach(status -> {
                 System.out.println("=============");
                 System.out.println(status.getAccount().getDisplayName());
